refactor(node): move _options.ts to node/internal
